+++ .gitignore
@@ -127,3 +127,6 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+log
+*.LAPTOP-TAPEGVFU
