+++ utils/train.py
@@ -1,67 +1,237 @@
 import time
 import logging
-from data_loader import get_dataloaders
-
+import os
 from models.TCFormer import tcformer
 
+import numpy as np
 import torch
 
-from utils.optimizer import build_optimizer
+from utils.evaluate import accuracy, get_final_preds
+from utils.loss import JointsMSELoss
+from utils.transforms import flip_back
+from utils.vis import save_debug_images
+def train(args,train_loader,model,optimizer, epoch,loss_func, log_writer):
+    batch_time = AverageMeter()
+    data_time = AverageMeter()
+    losses = AverageMeter()
+    acc = AverageMeter()
 
-from utils.validate import validate
-def train(args,model):
-    # 1. load data
-    train_dataloader, val_dataloader = get_dataloaders(args)
+    model.train()
 
-    # 2. build model and optimizers
-    model = tcformer()
-    checkpoint = torch.load(args.ckpt_file, map_location='cpu')
-    model.load_state_dict(checkpoint['model_state_dict'])
-    optimizer, scheduler = build_optimizer(args, model)
+    end = time.time()
+    for i, (x, target, target_weight, meta) in enumerate(train_loader):
+        # measure data loading time
+        data_time.update(time.time() - end)
 
-    if torch.cuda.is_available():
-        model = torch.nn.parallel.DataParallel(model.cuda())
-    optimizer.load_state_dict(checkpoint['optimizer'])
-    scheduler.load_state_dict(checkpoint['scheduler'])
+        target = target.cuda(non_blocking=True)
+        target_weight = target_weight.cuda(non_blocking=True)
+      
+            # compute output
+        outputs = model(x,target,target_weight)
+
+        if isinstance(outputs, list):
+                loss = loss_func(outputs[0], target, target_weight)
+                for output in outputs[1:]:
+                    loss += loss_func(output, target, target_weight)
+        else:
+                output = outputs
+                loss = loss_func(output, target, target_weight)
+
+        loss = loss_func(output, target, target_weight)
+        _, avg_acc, cnt, pred = accuracy(output.cpu().detach().numpy(),
+                                             target.cpu().numpy())
+        losses.update(loss)
+        acc.update(avg_acc)
+            # compute gradient and do update step
+        optimizer.zero_grad()
+        loss.backward()
+        optimizer.step()
+        # measure elapsed time
+        batch_time.update(time.time() - end)
+        end = time.time()
+
+        if i % args.print_steps == 0:
+            msg = 'Epoch: [{0}][{1}/{2}]\t' \
+                  'Time {batch_time.val:.3f}s ({batch_time.avg:.3f}s)\t' \
+                  'Speed {speed:.1f} samples/s\t' \
+                  'Data {data_time.val:.3f}s ({data_time.avg:.3f}s)\t' \
+                  'Loss {loss.val:.5f} ({loss.avg:.5f})\t' \
+                  'Accuracy {acc.val:.3f} ({acc.avg:.3f})'.format(
+                      epoch, i, len(train_loader), batch_time=batch_time,
+                      speed=x.size(0)/batch_time.val,
+                      data_time=data_time, loss=losses, acc=acc)
+            logging.info(msg)
+
+            writer = log_writer['writer']
+            global_steps = log_writer['train_global_steps']
+            writer.add_scalar('train_loss', losses.val, global_steps)
+            writer.add_scalar('train_acc', acc.val, global_steps)
+            log_writer['train_global_steps'] = global_steps + 1
+
+            prefix = '{}_{}'.format(os.path.join(args.debug_dir, 'train'), i)
+            save_debug_images(args, x, meta, target, pred*4, output,
+                              prefix)
+
+
+
+class AverageMeter(object):
+    """Computes and stores the average and current value"""
+    def __init__(self):
+        self.reset()
+
+    def reset(self):
+        self.val = 0
+        self.avg = 0
+        self.sum = 0
+        self.count = 0
+
+    def update(self, val, n=1):
+        self.val = val
+        self.sum += val * n
+        self.count += n
+        self.avg = self.sum / self.count if self.count != 0 else 0
+
+
+def validate(args, val_loader, val_dataset, model, loss_func, writer_dict=None):
+    batch_time = AverageMeter()
+    losses = AverageMeter()
+    acc = AverageMeter()
+
+    # switch to evaluate mode
     model.eval()
-    patience = args.patience
-    epoch_counter = 0
-    f1_scores = []
-    # 3. training
-    step = checkpoint['step']
-    best_score = args.best_score
-    start_time = time.time()
-    num_total_steps = len(train_dataloader) * args.max_epochs
-    for epoch in range(checkpoint['epoch']+1,args.max_epochs):
-        if epoch_counter > patience:
-            logging.info(f"Epoch {epoch} step {step} ---------------early stopping.-------------")
-            break
-        for batch in train_dataloader:
-            model.train()
-            
-            loss, accuracy, _, _ = model(batch)
-            loss = loss.mean()
-            accuracy = accuracy.mean()
-            loss.backward()
-            optimizer.step()
-            optimizer.zero_grad()
-            scheduler.step()
 
-            step += 1
-            if step % args.print_steps == 0:
-                time_per_step = (time.time() - start_time) / max(1, step)
-                remaining_time = time_per_step * (num_total_steps - step)
-                remaining_time = time.strftime('%H:%M:%S', time.gmtime(remaining_time))
-                logging.info(f"Epoch {epoch} step {step} eta {remaining_time}: loss {loss:.3f}, accuracy {accuracy:.3f}")
+    num_samples = len(val_dataset)
+    all_preds = np.zeros(
+        (num_samples, args.num_keypoints, 3),
+        dtype=np.float32
+    )
+    all_boxes = np.zeros((num_samples, 6))
+    image_path = []
+    filenames = []
+    imgnums = []
+    idx = 0
+    with torch.no_grad():
+        end = time.time()
+        for i, (x, target, target_weight, meta) in enumerate(val_loader):
+            # measure data loading time
 
-        # 4. validation
+            target = target.cuda(non_blocking=True)
+            target_weight = target_weight.cuda(non_blocking=True)
+            num_images = x.size(0)
+            # compute output
+            outputs = model(x,target,target_weight)
 
-        loss = validate(model, val_dataloader)
-        results = {k: round(v, 4) for k, v in results.items()}
-        logging.info(f"Epoch {epoch} step {step}: loss {loss:.3f}")
+            if isinstance(outputs, list):
+                loss = loss_func(outputs[0], target, target_weight)
+                for output in outputs[1:]:
+                    loss += loss_func(output, target, target_weight)
+            else:
+                output = outputs
+                loss = loss_func(output, target, target_weight)
 
-        # 5. save checkpoint
+            loss = loss_func(output, target, target_weight)
+            # measure accuracy and record loss
+            losses.update(loss.item(), x.size(0))
 
-        state_dict = model.module.state_dict() if args.device == 'cuda' else model.state_dict()
-        torch.save({'epoch': epoch, "step":step,'model_state_dict': state_dict,'optimizer':optimizer.state_dict(),'scheduler':scheduler.state_dict()},
-                       f'{args.savedmodel_path}/model_epoch_{epoch}.bin')
+            _, avg_acc, cnt, pred = accuracy(output.detach().cpu().numpy(),
+                                            target.detach().cpu().numpy())
+            acc.update(avg_acc, cnt)
+
+            # measure elapsed time
+            batch_time.update(time.time() - end)
+            end = time.time()
+
+            c = meta['center'].numpy()
+            s = meta['scale'].numpy()
+            score = meta['score'].numpy()
+
+            preds, maxvals = get_final_preds(
+                args, output.clone().cpu().numpy(), c, s)
+
+            all_preds[idx:idx + num_images, :, 0:2] = preds[:, :, 0:2]
+            all_preds[idx:idx + num_images, :, 2:3] = maxvals
+            # double check this all_boxes parts
+            all_boxes[idx:idx + num_images, 0:2] = c[:, 0:2]
+            all_boxes[idx:idx + num_images, 2:4] = s[:, 0:2]
+            all_boxes[idx:idx + num_images, 4] = np.prod(s*200, 1)
+            all_boxes[idx:idx + num_images, 5] = score
+            image_path.extend(meta['image'])
+
+            idx += num_images
+
+            if i % args.print_steps == 0:
+                msg = 'Test: [{0}/{1}]\t' \
+                      'Time {batch_time.val:.3f} ({batch_time.avg:.3f})\t' \
+                      'Loss {loss.val:.4f} ({loss.avg:.4f})\t' \
+                      'Accuracy {acc.val:.3f} ({acc.avg:.3f})'.format(
+                          i, len(val_loader), batch_time=batch_time,
+                          loss=losses, acc=acc)
+                logging.info(msg)
+
+                prefix = '{}_{}'.format(
+                    os.path.join(args.output_dir, 'val'), i
+                )
+                save_debug_images(args, x, meta, target, pred*4, output,
+                                  prefix)
+
+        name_values, perf_indicator = val_dataset.evaluate(
+            args, all_preds, args.output_dir, all_boxes, image_path,
+            filenames, imgnums
+        )
+
+        model_name = args.model_name
+        if isinstance(name_values, list):
+            for name_value in name_values:
+                _print_name_value(name_value, model_name)
+        else:
+            _print_name_value(name_values, model_name)
+
+        if writer_dict:
+            writer = writer_dict['writer']
+            global_steps = writer_dict['valid_global_steps']
+            writer.add_scalar(
+                'valid_loss',
+                losses.avg,
+                global_steps
+            )
+            writer.add_scalar(
+                'valid_acc',
+                acc.avg,
+                global_steps
+            )
+            if isinstance(name_values, list):
+                for name_value in name_values:
+                    writer.add_scalars(
+                        'valid',
+                        dict(name_value),
+                        global_steps
+                    )
+            else:
+                writer.add_scalars(
+                    'valid',
+                    dict(name_values),
+                    global_steps
+                )
+            writer_dict['valid_global_steps'] = global_steps + 1
+
+    return perf_indicator,loss
+
+# markdown format output
+def _print_name_value(name_value, full_arch_name):
+    names = name_value.keys()
+    values = name_value.values()
+    num_values = len(name_value)
+    logging.info(
+        '| Arch ' +
+        ' '.join(['| {}'.format(name) for name in names]) +
+        ' |'
+    )
+    logging.info('|---' * (num_values+1) + '|')
+
+    if len(full_arch_name) > 15:
+        full_arch_name = full_arch_name[:8] + '...'
+    logging.info(
+        '| ' + full_arch_name + ' ' +
+        ' '.join(['| {:.3f}'.format(value) for value in values]) +
+         ' |'
+    )
